% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{funSetHet}
\alias{funSetHet}
\title{Simulate heterogeneous exposure effects for scenario SetHet}
\usage{
funSetHet(X, Z, s, het.exposure.strength, snr, n_het_effects, signs)
}
\arguments{
\item{X}{Vector of the exposure variable.}

\item{Z}{Matrix of covariates.}

\item{s}{Vector of binary variable (e.g., Sex) of length nrow(Z).}

\item{het.exposure.strength}{Numeric. Strength of effect modification.}

\item{snr}{Numeric. Signal-to-noise ratio multiplier.}

\item{n_het_effects}{Integer. Number of covariates with heterogeneous effects.}

\item{signs}{Numeric vector of effect signs (+1/-1) corresponding to columns of Z.}
}
\value{
A list with components:
\describe{
\item{effect}{Numeric vector of simulated heterogeneous effects.}
\item{coefs}{Numeric vector of true heterogeneous coefficients.}
\item{n_het_effects}{Number of heterogeneous effects simulated.}
}
}
\description{
Generates heterogeneous exposure effects for covariates and a binary variable (s).
}
