% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_results}
\alias{plot_results}
\title{Plot Stability Selection Results with Cleaned Variable Names}
\usage{
plot_results(results, remove_patterns = NULL, ...)
}
\arguments{
\item{results}{A \code{stabsel} object returned from a stability
selection procedure (e.g., the output of \code{CLogitBoostHEE}).}

\item{remove_patterns}{A \code{list} object that is not mandatory but can be
specified to remove certain parts of baselearner nomenclature.}
}
\value{
Invisibly returns \code{NULL}. Generates a plot of the stability
selection results.
}
\description{
Plots a \code{stabsel} object while cleaning variable names by removing
technical annotations such as \code{df = 1}, \code{intercept = FALSE/TRUE},
and \code{center = TRUE/FALSE}. The plot retains the original \code{stabsel}
gradient bars and layout, with a customizable title.
}
\examples{
\dontrun{
# Suppose stab_model is a stabsel object from your analysis
plot_results(stab_model)
}

}
