% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{sim}
\alias{sim}
\title{Simulated dataset for examples}
\format{
A data frame with 4000 rows (1000 cases and 3000 controls) and 14 variables:
\describe{
\item{X}{Exposure}
\item{Z1-Z10}{Covariables}
\item{y}{Outcome}
\item{s}{Matching Variable}
\item{strata}{Strata ID}
...
}
}
\source{
Simulated
}
\usage{
sim
}
\description{
This dataset contains simulated data used for examples and testing.
}
\keyword{datasets}
